Loading for all pages integrated
