+++ website/src/components/Header.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-
+//className="active"
 function Header() {
   return (
 <header id="header" className="header d-flex align-items-center sticky-top">
@@ -12,7 +12,7 @@ function Header() {
 
   <nav id="navmenu" className="navmenu">
     <ul>
-      <li><Link  className="active" to="/">Home</Link></li>
+      <li><Link to="/">Home</Link></li>
       <li><Link to="/about">About</Link></li>
       <li><Link to="/services">Services</Link></li>
       <li><Link to="/contact">Contact</Link></li>
