+++ website/src/pages/Contact.jsx
@@ -1,10 +1,13 @@
-import React from 'react';
+import React, { useState, useEffect } from 'react';
+import Loading from './../components/Loading';
 
 function Contact() {
+ 
   return (
-
-            <p>Contact</p>
-
+  <>
+    <Loading></Loading>
+    <p>Contact</p>
+  </>
   );
 }
